requirements.txt
langchain
openai==0.27.4
huggingface_hub
python-dotenv
streamlit
